Test_Run/requirements.txt
#Stuff here yet!

sqlalchemy
psycopg2-binary
pandas
streamlit
numpy


